add dark mode switch to menu dropdown
